chore: change 'changeAnimal' fun position to app.component.ts
