+++ apps/give-me-cute-animals/src/app/app.component.html
@@ -1,4 +1,4 @@
-<app-header [(animal)]="animal"></app-header>
+<app-header [animal]="animal" (changeAnimalEmitter)="changeAnimal()"></app-header>
 <main>
 <app-home [animal]="animal"></app-home>
 </main>
