+++ apps/give-me-cute-animals/src/app/app.component.ts
@@ -1,6 +1,7 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { Component } from '@angular/core';
 
 import { Animal } from 'src/core/models/animal.model';
+import { AnimalService } from 'src/core/services/animal.service';
 
 @Component({
   selector: 'app-root',
@@ -8,5 +9,38 @@ import { Animal } from 'src/core/models/animal.model';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  @Input() animal?: Animal;
+  animal?: Animal;
+
+  constructor(private animalService: AnimalService) { }
+
+  ngOnInit(): void {
+    this.changeAnimal();
+  }
+
+  changeAnimal(): void {
+    const randomNumber = Math.floor(Math.random() * 3) + 1;
+
+    switch (randomNumber) {
+      case 1:
+        this.animalService.getRandomCat().subscribe((data:Animal) => {
+          this.animal = data;
+        })
+        break;
+      case 2:
+        this.animalService.getRandomDog().subscribe((data:Animal) => {
+          this.animal = data;
+        })
+        break;
+      case 3:
+        this.animalService.getRandomFox().subscribe((data:Animal) => {
+          this.animal = data;
+        })
+        break;
+      default:
+        this.animalService.getRandomCat().subscribe((data:Animal) => {
+          this.animal = data;
+        })
+        break;
+    }
+  }
 }
