+++ apps/give-me-cute-animals/src/app/shared/components/header/header.component.html
@@ -4,8 +4,8 @@
             <h1>give me cute animals</h1>
         </div>
         <div class="col-12 col-lg-9 text-center text-lg-right">
-            <a (click)="getAnimal()" class="d-none d-lg-inline-block btn btn-link text-white">show me another one</a>
-            <a (click)="getAnimal()" class="d-inline-block d-lg-none btn btn-danger my-3 text-white">show me another one</a>
+            <a (click)="changeAnimal()" class="d-none d-lg-inline-block btn btn-link text-white">show me another one</a>
+            <a (click)="changeAnimal()" class="d-inline-block d-lg-none btn btn-danger my-3 text-white">show me another one</a>
         </div>    
     </div>
 </nav>
